requirements.txt
asgiref>=3.5.2
Django>=4.0.6
sqlparse==0.4.2
uWSGI>=2.0.18,<2.1
python-dotenv==0.20.0
djangorestframework~=3.13.1
